making WED components compile on linux
this commit also includes changes which allows compiling on x86_64 platform.
I specifically exchanged int with long in some functions where pointers
come into play (i.e. BroadcastMessage() and such). since long is 32 bit wide
under windows with msvc it will only work under linux when compiling with gcc.
32 bit builds should not be affected anyway by this change. Most changes in
this commit are case sensitivity fixes.

+++ src/GUI/GUI_Broadcaster.cpp
@@ -38,7 +38,7 @@ GUI_Broadcaster::~GUI_Broadcaster()
 	}
 }
 	
-void	GUI_Broadcaster::BroadcastMessage(int inMsg, int inParam)
+void	GUI_Broadcaster::BroadcastMessage(long inMsg, long inParam)
 {
 	// BEN SEZ: for now - just make sure that a suicidal listener doesn't
 	// @#$#@ us up by editing the list while we're on it!
