+++ src/GUI/GUI_ScrollerPane.h
@@ -101,8 +101,8 @@ public:
 	// From GUI_Listener
 	virtual	void	ReceiveMessage(
 							GUI_Broadcaster *		inSrc,
-							int						inMsg,
-							int						inParam);
+							long					inMsg,
+							long					inParam);
