+++ src/GUI/GUI_SimpleScroller.h
@@ -50,8 +50,8 @@ public:
 
 	virtual	void	ReceiveMessage(
 							GUI_Broadcaster *		inSrc,
-							int						inMsg,
-							int						inParam);
+							long					inMsg,
+							long					inParam);
 private:
 			void	AlignContents(void);
