+++ src/GUI/GUI_TabPane.h
@@ -48,8 +48,8 @@ public:
 
 	virtual	void			ReceiveMessage(
 									GUI_Broadcaster *		inSrc,
-									int						inMsg,
-									int						inParam);
+									long					inMsg,
+									long					inParam);
 			
 private:
