+++ src/GUI/GUI_Table.cpp
@@ -29,7 +29,7 @@
 #if APL
 	#include <OpenGL/gl.h>
 #else	
-	#include <gl/gl.h>
+	#include <GL/gl.h>
 #endif
 
 #define	AUTOSCROLL_DIST	15
@@ -487,8 +487,8 @@ void		GUI_Table::AlignContents(void)
 
 void		GUI_Table::ReceiveMessage(
 				GUI_Broadcaster *		inSrc,
-				int						inMsg,
-				int						inParam)
+				long					inMsg,
+				long					inParam)
 {
 	switch(inMsg) {
 	case GUI_TABLE_SHAPE_RESIZED:
@@ -772,8 +772,8 @@ int		GUI_Header::GetHelpTip(int x, int y, int tip_bounds[4], string& tip)
 
 void		GUI_Header::ReceiveMessage(
 				GUI_Broadcaster *		inSrc,
-				int						inMsg,
-				int						inParam)
+				long					inMsg,
+				long					inParam)
 {
 	switch(inMsg) {
 	case GUI_TABLE_SHAPE_RESIZED:
@@ -963,8 +963,8 @@ int		GUI_Side::GetHelpTip(int x, int y, int tip_bounds[4], string& tip)
 
 void		GUI_Side::ReceiveMessage(
 				GUI_Broadcaster *		inSrc,
-				int						inMsg,
-				int						inParam)
+				long					inMsg,
+				long					inParam)
 {
 	switch(inMsg) {
 	case GUI_TABLE_SHAPE_RESIZED:
