+++ src/WEDMap/WED_MapPane.h
@@ -74,8 +74,8 @@ public:
 
 	virtual	void	ReceiveMessage(
 							GUI_Broadcaster *		inSrc,
-							int						inMsg,
-							int						inParam);
+							long					inMsg,
+							long					inParam);
 			
 private:
