+++ src/WEDProperties/WED_PropertyTable.cpp
@@ -823,8 +823,8 @@ int			WED_PropertyTable::GetRowCount(void)
 
 void	WED_PropertyTable::ReceiveMessage(
 							GUI_Broadcaster *		inSrc,
-							int						inMsg,
-							int						inParam)
+							long					inMsg,
+							long					inParam)
 {
 //	if (inMsg == msg_SelectionChanged)		BroadcastMessage(GUI_TABLE_CONTENT_CHANGED,0);
 	if (inMsg == msg_ArchiveChanged)
