+++ src/WEDWindows/WED_DocumentWindow.h
@@ -54,8 +54,8 @@ public:
 
 	virtual	void	ReceiveMessage(
 							GUI_Broadcaster *		inSrc,
-							int						inMsg,
-							int						inParam);
+							long					inMsg,
+							long					inParam);
 
 	virtual	bool	Closed(void);
